reflex_frontend/requirements.txt
reflex>=0.7.13,<0.8.0
python-dotenv
pandas
plotly
httpx
pytest
playwright